feat: BeverageDomain에 History 모델 및 API 로직 통합
