+++ HumanNeverDie/Projects/Domain/BaseDomain/Project.swift
@@ -5,8 +5,7 @@ let project = Project.makeModule(
     name: Modules.BaseDomain.name,
     product: .framework,
     dependencies: [.domain(module: .Beverage),
-        .domain(module: .User),
-        .domain(module: .History)
+        .domain(module: .User)
     ],
     hasTests: false
 )
