+++ HumanNeverDie/Projects/Domain/BeverageDomain/Sources/UseCases/BeverageUseCase+Dependency.swift
@@ -28,6 +28,7 @@ private struct MockBeverageUseCase: BeverageUseCaseProtocol {
   func getBeverageCount() async throws -> BeverageCount { BeverageCount.mock() }
   func getBeverageList(cursor: String?) async throws -> BeverageList { BeverageList.mock() }
   func getBeverageDetail(productID: String) async throws -> BeverageDetail { BeverageDetail.mock() }
+  func getBeverageMonthCalender(dateInWeek: String) async throws -> [BeverageCalendar] { BeverageCalendar.mock() }
   func likeBeverage(productID: String) async throws -> BeverageLike { BeverageLike.mock() }
   func unLikeBeverage(productID: String) async throws -> BeverageLike { BeverageLike.mock() }
   func searchBeverage(keyword: String) async throws -> BeverageList { BeverageList.mock() }
