--- HumanNeverDie/Projects/Domain/HistoryDomain/Project.swift
@@ -1,16 +1,0 @@
-//
-//  Project.swift
-//  HistoryDomain
-//
-//  Created by Seulki Lee on 2025/07/30.
-//
-
-import ProjectDescription
-import ProjectDescriptionHelpers
-
-let project = Project.makeDomain(
-    name: Modules.Domain.History.rawValue,
-    dependencies: [
-        .shared
-    ]
-)
