--- HumanNeverDie/Projects/Domain/HistoryDomain/Sources/Entity/DailyHistoryData.swift
@@ -1,8 +1,0 @@
-//
-//  DailyHistoryData.swift
-//  HistoryDomain
-//
-//  Created by Seulki Lee on 8/1/25.
-//
-
-import Foundation
