--- HumanNeverDie/Projects/Domain/HistoryDomain/Tests/Sources/HistoryUseCaseTest.swift
@@ -1,8 +1,0 @@
-//
-//  HistoryUseCaseTest.swift
-//  HistoryDomainTest
-//
-//  Created by Seulki Lee on 2025/07/30.
-//
-
-@testable import HistoryDomain
